index: Extract abstract IndexBase class
